requirements.txt
django==2.2.10
dj-database-url==0.5.0
django-heroku
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
django-extensions==2.2.1
django-allauth==0.36.0
django-rest-auth==0.9.3
django-debug-toolbar==2.0
django-environ==0.4.4
django-anymail==7.0.0
django-phonenumber-field==4.0.0
psycopg2-binary==2.8.3
django-cors-headers==2.2.0
python-decouple==3.1
phonenumbers==8.12.1
ipython
gunicorn==19.9.0
boto3==1.7.58
django-storages==1.6.6
django-redis==4.9.0
django-q==1.0.2
WeasyPrint==50
pylint-django==2.0.11
requests==2.22.0
sentry-sdk
django-q[sentry]
django-money==0.15
